Create side-by-side project directories to enable compilation with multiple versions of Delphi.
Hope I'm right that the .DPR is the main project file for Delphi 6, otherwise there's no point in copy/pasting it.  Can't see any other files that looks like a Delphi 6 main project file..

I've added helpers.pas to the Delphi 10 project because it wouldn't compile without it.  It's not in the Delphi 6 project, so I assume everything just magically works there?