Bugfix: IPC messages were being sent to the Application, but the handlers for these messages are in the Main Form.  The bug is a regression caused by r321.  Fixes issue #1629603 ("export to another host doesn't work").
We now store the main form's handle in the window list, though it would seem nicer to store the top-most handle (the application's handle).  The reason is that it's considerably easier to find the application's handle given the main form's handle than the other way around.

As usual, thanks to Peter Below and TeamB for their very helpful postings on Usenet ;-).