Avoid checking access to privileges database at startup, better do it on demand, when the user clicks the button for the user manager.
Taking into account that the user may not have the "SHOW DATABASE" privilege, but nevertheless could have access to the privileges tables. HasAccessToDB() was broken and useless in this way as seen in the forum: http://www.heidisql.com/forum/viewtopic.php?t=174

Adds the "DisplayErrors"-parameter to the ExecXXQuery-methods so we can suppress the MessageDlg if we want to show another error message.