* Bugfix: Allow editing the 'fallback' password and privileges.
Note on verbiage:  I chose "Everybody" rather than "Everyone" or "Fallback" as the pretty name for this account.

"Fallback" just seemed like it was misplaced in the user dropdown list.

"Everyone" seems to mean individuals in a group, while "everybody" seems to mean the group as a whole.  Thus "everyone" is a bit more personal.

This could convey the notion that you need to have a personal account before these privileges will apply to you, which is not the case.  Anyone that logs in with a non-existent user name will get these privileges.

Therefore, "Everybody".  It hopefully denotes that the privileges are given to everyone with a personal account, PLUS people authenticating with a rubbish user name (and the password, if any).

This last category of logins are commonly called "anonymous users" in the MySQL documentation.  "Anonymous" does not quite fit this account however, because the privileges given also apply to everyone else.

* Bugfix: Copy + paste same text in the username field caused a warning.